redo webpage, update streaming config
This commit redoes the site since the previous template we were using
didn't have an explicit license and was legally questionable. It's
redone from scratch now in Bootstrap. This commit adds a page with shiny
transitions and the video element.

This commit also makes some important changes to the streaming config,
namely it has changes that should decrease lag, especially when the
server is left running.